refactor/1179 - adds SettingsController
