+++ src/assets/scripts/client/ui/UiController.js
@@ -1,35 +1,13 @@
 import $ from 'jquery';
-import _forEach from 'lodash/forEach';
-import _isNaN from 'lodash/isNaN';
 import _keys from 'lodash/keys';
 import AirportController from '../airport/AirportController';
 import EventBus from '../lib/EventBus';
 import GameController from '../game/GameController';
+import SettingsController from './SettingsController';
 import { speech_toggle } from '../speech';
 import { EVENT } from '../constants/eventNames';
 import { INVALID_NUMBER } from '../constants/globalConstants';
 import { SELECTORS } from '../constants/selectors';
-
-/**
- * @property UI_SETTINGS_MODAL_TEMPLATE
- * @type {string}
- * @final
- */
-const UI_SETTINGS_MODAL_TEMPLATE = '<div class="option-dialog"></div>';
-
-/**
- * @property UI_OPTION_CONTAINER_TEMPLATE
- * @type {string}
- * @final
- */
-const UI_OPTION_CONTAINER_TEMPLATE = '<div class="option"></div>';
-
-/**
- * @property UI_OPTION_SELECTOR_TEMPLATE
- * @type {string}
- * @final
- */
-const UI_OPTION_SELECTOR_TEMPLATE = '<span class="option-type-select"></span>';
 
 /**
  * @class UiController
@@ -68,6 +46,8 @@ class UiController {
      * @param $element {jQuery Element}
      */
     init($element) {
+        this.settingsController = new SettingsController($element);
+
         this.$element = $element;
 
         this.$airportList = this.$element.find(SELECTORS.DOM_SELECTORS.AIRPORT_LIST);
@@ -182,21 +162,6 @@ class UiController {
      */
     ui_init() {
         this.$fastForwards.prop('title', 'Set time warp to 2');
-
-        const $options = $(UI_SETTINGS_MODAL_TEMPLATE);
-        const descriptions = GameController.game.option.getDescriptions();
-
-        _forEach(descriptions, (opt) => {
-            if (opt.type !== 'select') {
-                return;
-            }
-
-            const $container = this._buildOptionTemplate(opt);
-            $options.append($container);
-        });
-
-        $('body').append($options);
-
         // TODO: Make the options dialog findable by ID, not just by class
         this.$optionsDialog = this.$element.find(SELECTORS.DOM_SELECTORS.OPTIONS_DIALOG);
     }
@@ -254,68 +219,6 @@ class UiController {
      */
     isTutorialDialogOpen() {
         return this.$tutorialDialog.hasClass(SELECTORS.CLASSNAMES.OPEN);
-    }
-
-    /**
-     * Build the html for a game option and its cooresponding value elements.
-     *
-     * @for UiController
-     * @method _buildOptionTemplate
-     * @param option {object}
-     * @return $container {jquery Element}
-     * @private
-     */
-    _buildOptionTemplate(option) {
-        const $container = $(UI_OPTION_CONTAINER_TEMPLATE);
-        $container.append(`<span class="option-description">${option.description}</span>`);
-
-        const $optionSelector = $(UI_OPTION_SELECTOR_TEMPLATE);
-        const $selector = $(`<select name="${option.name}"></select>`);
-        const selectedOption = GameController.game.option.getOptionByName(option.name);
-
-        // this could me done with a _map(), but verbosity here makes the code easier to read
-        for (let i = 0; i < option.optionList.length; i++) {
-            const $optionSelectTempalate = this._buildOptionSelectTemplate(option.optionList[i], selectedOption);
-
-            $selector.append($optionSelectTempalate);
-        }
-
-        // TODO: this should be moved to proper event handler method and only assigned here.
-        $selector.change((event) => {
-            const $currentTarget = $(event.currentTarget);
-
-            GameController.game.option.setOptionByName($currentTarget.attr('name'), $currentTarget.val());
-        });
-
-        $optionSelector.append($selector);
-        $container.append($optionSelector);
-
-        return $container;
-    }
-
-    /**
-     * Build the html for a select option.
-     *
-     * @for UiController
-     * @method _buildOptionTemplate
-     * @param optionData {array<string>}
-     * @param selectedOption {string}
-     * @return optionSelectTempalate {HTML Element}
-     * @private
-     */
-    _buildOptionSelectTemplate(optionData, selectedOption) {
-        // the `selectedOption` coming in to this method will always be a string (due to existing api) but
-        // could contain valid numbers. here we test for valid number and build `parsedSelectedOption` accordingly.
-        const parsedSelectedOption = !_isNaN(parseFloat(selectedOption))
-            ? parseFloat(selectedOption)
-            : selectedOption;
-        let optionSelectTempalate = `<option value="${optionData.value}">${optionData.displayLabel}</option>`;
-
-        if (optionData.value === parsedSelectedOption) {
-            optionSelectTempalate = `<option value="${optionData.value}" selected>${optionData.displayLabel}</option>`;
-        }
-
-        return optionSelectTempalate;
     }
 
     /**
